feat(graph): amend commit
